chore: fix pg store and test coverage
